Switch to lucide and fix copyright symbol
Luckily for us, there's an extension of feather, called lucide, that has
a copyright symbol and works just as well, or so it seems. Consequently,
we switch out feather for lucide everywhere.

Also, the copyright symbol was horribly off-center, so I added a css
class to remedy this and fixed all the other symbols in the same stroke.