Change default colors and clean up css
The css was a bit messy, so I cleaned it up in part by introducing new
variables, changing some of the spacing, and changing capitalization.
Second, change some of the default colors; in particular exchange the
ghastly green that was the default for the dark theme for orange red,
thereby eliminating the risk for spontaneous ophthalmologic hemorrhage.