Refactor Readme
In particular, remove the images of archie and the link to the demo site
and replace it with a sad "not available" message. Furthermore, fix
several grammar and spelling errors and restructure the general sentence
structure to be less awkward. Finally, add a section detailing what the
differences between this theme and archie are, since they're obviously
(still) very similar in nature.